Update wdl tools 0.5 (#443)
Update to latest WDL tools (0.5.0):

Code reorg (e.g. wdlTools.util.Utils was split into FileUtils and SystemUtils) and some functions renamed (e.g. exprToString -> prettyFormatExpr)
WDL Ints are now represented as Scala Longs
TypedAbstractSyntax was simplified and some classes removed/unified
Eval API changed, and Context was removed in favor of the more generic Bindings class

+++ src/main/scala/dx/compiler/IR.scala
@@ -44,7 +44,7 @@ object IR {
   sealed abstract class MetaValue
   final case object MetaValueNull extends MetaValue
   final case class MetaValueBoolean(value: Boolean) extends MetaValue
-  final case class MetaValueInt(value: Int) extends MetaValue
+  final case class MetaValueInt(value: Long) extends MetaValue
   final case class MetaValueFloat(value: Double) extends MetaValue
   final case class MetaValueString(value: String) extends MetaValue
   final case class MetaValueObject(value: Map[String, MetaValue]) extends MetaValue
@@ -92,13 +92,13 @@ object IR {
   val ALL_KEY = "All"
 
   sealed abstract class RuntimeHint
-  final case class RuntimeHintRestart(max: Option[Int] = None,
-                                      default: Option[Int] = None,
-                                      errors: Option[Map[String, Int]] = None)
+  final case class RuntimeHintRestart(max: Option[Long] = None,
+                                      default: Option[Long] = None,
+                                      errors: Option[Map[String, Long]] = None)
       extends RuntimeHint
-  final case class RuntimeHintTimeout(days: Option[Int] = None,
-                                      hours: Option[Int] = None,
-                                      minutes: Option[Int] = None)
+  final case class RuntimeHintTimeout(days: Option[Long] = None,
+                                      hours: Option[Long] = None,
+                                      minutes: Option[Long] = None)
       extends RuntimeHint
   final case class RuntimeHintIgnoreReuse(value: Boolean) extends RuntimeHint
   final case class RuntimeHintAccess(network: Option[Vector[String]] = None,
@@ -191,7 +191,7 @@ object IR {
   **/
   sealed abstract class ChoiceRepr
   final case class ChoiceReprString(value: String) extends ChoiceRepr
-  final case class ChoiceReprInteger(value: Int) extends ChoiceRepr
+  final case class ChoiceReprInteger(value: Long) extends ChoiceRepr
   final case class ChoiceReprFloat(value: Double) extends ChoiceRepr
   final case class ChoiceReprBoolean(value: Boolean) extends ChoiceRepr
   final case class ChoiceReprFile(value: String, name: Option[String]) extends ChoiceRepr
@@ -214,7 +214,7 @@ object IR {
   **/
   sealed abstract class SuggestionRepr
   sealed case class SuggestionReprString(value: String) extends SuggestionRepr
-  sealed case class SuggestionReprInteger(value: Int) extends SuggestionRepr
+  sealed case class SuggestionReprInteger(value: Long) extends SuggestionRepr
   sealed case class SuggestionReprFloat(value: Double) extends SuggestionRepr
   sealed case class SuggestionReprBoolean(value: Boolean) extends SuggestionRepr
   sealed case class SuggestionReprFile(
@@ -246,7 +246,7 @@ object IR {
   **/
   sealed abstract class DefaultRepr
   final case class DefaultReprString(value: String) extends DefaultRepr
-  final case class DefaultReprInteger(value: Int) extends DefaultRepr
+  final case class DefaultReprInteger(value: Long) extends DefaultRepr
   final case class DefaultReprFloat(value: Double) extends DefaultRepr
   final case class DefaultReprBoolean(value: Boolean) extends DefaultRepr
   final case class DefaultReprFile(value: String) extends DefaultRepr
@@ -306,9 +306,9 @@ object IR {
   case object InstanceTypeDefault extends InstanceType
   case class InstanceTypeConst(
       dxInstanceType: Option[String],
-      memoryMB: Option[Int],
-      diskGB: Option[Int],
-      cpu: Option[Int],
+      memoryMB: Option[Long],
+      diskGB: Option[Long],
+      cpu: Option[Long],
       gpu: Option[Boolean]
   ) extends InstanceType
   case object InstanceTypeRuntime extends InstanceType
